添加尽力而为放入策略

+++ SDK/SDK_C++/CodeCraft-2022/src/include/client_node.h
@@ -38,7 +38,7 @@ class ClientNode {
   };
 
   int GetAvailableEdgeNodeNum() {
-    return available_edgenode.size();
+    return (int) available_edgenode.size();
   };
 
   std::string ToString();
